Fix and Update Unreal Plugin for UE 5 (#49)
* Make Unreal plugin compilable
* Added get and set var for runtime
* Change BeginPlay order to load story first
* reset tags for unreal getline to keep tags line dependent
* Add *.ink loading support for unreal
* Fixing SetGlobalVariable to support Strings
* Add fallback for external functions
    https://github.com/inkle/ink/blob/master/Documentation/RunningYourInk.md#fallbacks-for-external-functions
    it is not possible to implement a dummy/error function for each external
    functions. It must have been the same name and same number of arguments
    and will be executed if the external function was not bind when needed.
    IF the external functions will get bind later, then it will be called
    the next time.
    Introduces new flag: "FALLBACK_FUNCTION" to add a new execution
    dimension
    Introduce new value type: `ex_fn_not_found` to signal failed external
    execution
* Make ex_fn_not_found check work on empty stack
* Fixed out-of-order definition
* Update Policy for cmake version 3.24.1
* Fixed wrong forget function of restorable_stack
* Add interface for creating and loading snapshots
    snapshot creation is only possible via runner because a runner without
    global state won't work. The fact that the runner also snapshots all
    other runner with the same global is a bit obscure and may should be
    changed, but since they are related \_o_o_/
* Can load global, start loading runner
* Fix access to union to conform with mvsc
* fixed dynamic tag list
* Fixes PRNG and implement seed interface for runner
* Make list available for modifications
* Access and modify lists, WITH memory leak
* choices tags not yet implemented in UE
* fixes wrong FString to UTF8 conversion
* Rename UChoice to UInkChoice
* Add Observers to inkcpp
* Add Doxygen Docu
* Delete temporary json ink file
* Bump InkcppVersion

+++ inkcpp_compiler/list_data.cpp
@@ -13,7 +13,7 @@ namespace ink::compiler::internal
 		_list_end.push_back(current_back);
 	}
 
-	void list_data::new_flag(const std::string& flag_name, int value)
+	void list_data::new_flag(const std::string& flag_name, size_t value)
 	{
 		while(_flag_names.size() < _current_list_start + value) {
 			_flag_names.push_back("");
@@ -23,6 +23,7 @@ namespace ink::compiler::internal
 		}
 		_flag_names[_current_list_start + value - 1] = flag_name;
 	}
+
 	std::vector<list_data::named_list_flag> list_data::get_flags() const {
 		std::vector<named_list_flag> result{};
 		size_t begin = 0;
